get report by admin modified

+++ schemas/order/order.schema.js
@@ -71,7 +71,7 @@ type ReportsOutput {
 type ReportData {
     restaurant: RestaurantData
     orders: [OrderData]
-    total: Int
+    total: Float
 }
 
 input OrderInput {
